services/workers/requirements.txt
boto3>=1.34
pydantic>=2.6
python-dotenv>=1.0
